# Improve performance of `vec_duplicate_detect()`