--- a/NEWS.md
+++ b/NEWS.md
@@ -1,5 +1,7 @@
 # vctrs (development version)
 
+* `vec_duplicate_detect()` is a bit faster when there are many unique values.
+
 * New `vec_rank()` to compute various types of sample ranks.
 
 * `vec_chop()` now materializes ALTREP vectors before chopping, which is more